Refactor discover url for admin toast
