+++ src/Plugins/AdminToast/GuessEditUrl.php
@@ -2,14 +2,15 @@
 
 declare(strict_types=1);
 
-namespace Thinktomorrow\Chief\Site;
+namespace Thinktomorrow\Chief\Plugins\AdminToast;
 
 use Thinktomorrow\Chief\Managers\Register\Registry;
 use Thinktomorrow\Chief\Shared\ModelReferences\ModelReference;
+use Thinktomorrow\Chief\Site\Urls;
 use Thinktomorrow\Chief\Site\Urls\UrlRecord;
 use Thinktomorrow\Chief\Site\Urls\UrlRecordNotFound;
 
-class AdminToast
+class GuessEditUrl
 {
     protected Registry $registry;
 
@@ -18,7 +19,7 @@ class AdminToast
         $this->registry = $registry;
     }
 
-    public function discoverEditUrl(string $path, string $locale): ?string
+    public function guessByPathAndLocale(string $path, string $locale): ?string
     {
         // Remove the locale segment if present - we assume the first segment is the locale
         if (0 === strpos($path, $locale.'/') || $path === $locale) {
